[REFACTOR] 유저 프로필 수정 시, AvatarId가 아닌 AvatarProfileId 변경하도록 변

+++ src/main/java/org/websoso/WSSServer/dto/user/UpdateMyProfileRequest.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.websoso.WSSServer.validation.NullAllowedNicknameConstraint;
 
 public record UpdateMyProfileRequest(
-        Byte avatarId,
+        Long avatarId,
 
         @NullAllowedNicknameConstraint
         String nickname,
